Make functions an seperate js files
